dir authentication renamed to registration
