+++ app/src/main/java/com/codecool/arinyu/myfirstandroidapp/MainActivity.java
@@ -14,7 +14,7 @@ import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import com.codecool.arinyu.myfirstandroidapp.authentication.RegistrationActivity;
+import com.codecool.arinyu.myfirstandroidapp.registration.RegistrationActivity;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
