+++ app/src/main/res/layout/activity_registration.xml
@@ -10,7 +10,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".authentication.RegistrationActivity">
+    tools:context=".registration.RegistrationActivity">
 
     <!-- Login progress -->
     <ProgressBar
